server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
pydantic
debugpy
newspaper3k